make files renaming
