+++ src/components/Header/Search/Search.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Button from "../../Button";
-import styles from "./styles.module.scss";
+import styles from "./Search.module.scss";
 
 const Search = (): React.ReactElement => {
   const searchMovies = () => console.log('searching ...');
@@ -10,7 +10,7 @@ const Search = (): React.ReactElement => {
       <h2 className={styles.h2}>Find your movie</h2>
       <div className={styles.searchWrapper}>
         <input className={styles.searchField} type="text" placeholder="What do you want to watch?" />
-        <Button text="Search" extraClass={styles.btn} onClick={searchMovies}/>
+        <Button text="Search" className={styles.btn} onClick={searchMovies}/>
       </div>
     </div>
   );
