+++ src/components/Logo/Logo.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./styles.module.scss";
+import styles from "./Logo.module.scss";
 
 const Logo = (): React.ReactElement => {
   return (
