+++ src/components/Main/Card/Card.module.scss
