+++ src/components/Main/DropdownSelect/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DropdownSelect";
