+++ src/components/Main/ResultsPanel/ResultsPanel.module.scss
