+++ src/components/Modal/Modal.tsx
@@ -1,7 +1,7 @@
 import classNames from "classnames";
 import React from "react";
 import Button from "../Button";
-import styles from "./styles.module.scss";
+import styles from "./Modal.module.scss";
 
 interface IModalProps {
   modalTitle: string;
@@ -82,7 +82,7 @@ class Modal extends React.Component<IModalProps, IModalState> {
         <form className={styles.modal} autoComplete="off">
           <Button
             text={""}
-            extraClass={styles.closeBtn}
+            className={styles.closeBtn}
             onClick={this.props.onClose}
           />
           <h2 className={styles.h2}>{this.props.modalTitle}</h2>
@@ -193,12 +193,12 @@ class Modal extends React.Component<IModalProps, IModalState> {
               </label>
               <Button
                 text={"Reset"}
-                extraClass={styles.resetBtn}
+                className={styles.resetBtn}
                 onClick={(e) => this.resetForm(e)}
               />
               <Button
                 text={"Submit"}
-                extraClass={styles.submitBtn}
+                className={styles.submitBtn}
                 onClick={this.props.onClose}
               />
             </>
@@ -210,7 +210,7 @@ class Modal extends React.Component<IModalProps, IModalState> {
               </p>
               <Button
                 text={"Confirm"}
-                extraClass={styles.submitBtn}
+                className={styles.submitBtn}
                 onClick={this.props.onClose}
               />
             </>
